working feedback with arrow
