+++ CMakeLists.txt
@@ -5,12 +5,12 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 SET(CMAKE_BUILD_TYPE RelWithDebInfo)
 
 find_package(catkin REQUIRED COMPONENTS 
-			 roscpp 
-			 std_msgs
-			 #dynamic_reconfigure
-			 rosneuro_msgs
-			 Eigen3
-			 )
+             roscpp 
+             std_msgs
+             dynamic_reconfigure
+             rosneuro_msgs
+             Eigen3
+             )
 
 find_library(NEURODRAW_LIBRARIES neurodraw REQUIRED)
 find_path(NEURODRAW_INCLUDE_DIRS neurodraw/Shape.h REQUIRED)
@@ -18,19 +18,19 @@ find_path(NEURODRAW_INCLUDE_DIRS neurodraw/Shape.h REQUIRED)
 find_library(NEUROCHRONO_LIBRARIES neurochrono REQUIRED)
 find_path(NEUROCHRONO_INCLUDE_DIRS neurochrono/Timer.h REQUIRED)
 
-#generate_dynamic_reconfigure_options(
-#	cfg/Wheel.cfg
-#)
+generate_dynamic_reconfigure_options(
+  cfg/CVSA.cfg
+)
 
 catkin_package(
   INCLUDE_DIRS 
-	include
+    include
   LIBRARIES 
-	${PROJECT_NAME}
+    ${PROJECT_NAME}
   CATKIN_DEPENDS
-  	roscpp
-	std_msgs
-	rosneuro_msgs
+    roscpp
+    std_msgs
+    rosneuro_msgs
   DEPENDS
 )
 
@@ -40,42 +40,43 @@ catkin_package(
 ###########
 
 include_directories(include 
-					${catkin_INCLUDE_DIRS}
-					${NEURODRAW_INCLUDE_DIRS}
-					${NEUROCHRONO_INCLUDE_DIRS}
-					${EIGEN3_INCLUDE_DIRS})
+                    ${catkin_INCLUDE_DIRS}
+                    ${NEURODRAW_INCLUDE_DIRS}
+                    ${NEUROCHRONO_INCLUDE_DIRS}
+                    ${EIGEN3_INCLUDE_DIRS})
 
 add_definitions(${EIGEN3_DEFINITIONS})
 
-add_library(${PROJECT_NAME}	
-			src/SingleArrow.cpp
-			src/TrainingArrow.cpp
-			#src/ControlWheel.cpp
-			src/TrialSequence.cpp
-			src/Autopilot.cpp
-			)
+add_library(${PROJECT_NAME}    
+            src/CVSA_layout.cpp
+            src/TrainingCVSA.cpp
+            #src/ControlWheel.cpp
+            src/TrialSequence.cpp
+            src/Autopilot.cpp
+            )
 
 target_link_libraries(${PROJECT_NAME} 
-					  ${catkin_LIBRARIES}
-					  ${NEURODRAW_LIBRARIES}
-					  ${NEUROCHRONO_LIBRARIES}
-					  ${Eigen3_LIBRARIES})
+                      ${catkin_LIBRARIES}
+                      ${NEURODRAW_LIBRARIES}
+                      ${NEUROCHRONO_LIBRARIES}
+                      ${Eigen3_LIBRARIES})
 
 add_dependencies(${PROJECT_NAME} 
-				 ${catkin_EXPORTED_TARGETS}
-				 #${PROJECT_NAME}_gencfg
+                 ${catkin_EXPORTED_TARGETS}
+                 ${PROJECT_NAME}_gencfg
 )
 
 #################
 ## Executables ##
 #################
-#add_executable(trainingwheel src/trainingwheel.cpp)
-#target_link_libraries(trainingwheel ${PROJECT_NAME}
-#								 ${catkin_LIBRARIES}
-#					  			 ${NEURODRAW_LIBRARIES}
-#					  			 ${NEUROCHRONO_LIBRARIES})
+add_executable(trainingCVSA_node src/trainingCVSA_node.cpp)
+target_link_libraries(trainingCVSA_node  ${PROJECT_NAME}
+                                    ${catkin_LIBRARIES}
+                                    ${NEURODRAW_LIBRARIES}
+                                    ${NEUROCHRONO_LIBRARIES}
+									                  ${Eigen3_LIBRARIES})
 #add_executable(controlwheel src/controlwheel.cpp)
 #target_link_libraries(controlwheel ${PROJECT_NAME}
-#								 ${catkin_LIBRARIES}
-#					  			 ${NEURODRAW_LIBRARIES}
-#					  			 ${NEUROCHRONO_LIBRARIES})
+#                                 ${catkin_LIBRARIES}
+#                                   ${NEURODRAW_LIBRARIES}
+#                                   ${NEUROCHRONO_LIBRARIES})
