+++ cfg/CVSA.cfg
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+PACKAGE = "feedback_cvsa"
+
+from dynamic_reconfigure.parameter_generator_catkin import *
+
+gen = ParameterGenerator()
+
+gen.add("leftbottom_threshold", double_t, 0, "Leftbottom threshold", 0.7, 0.5, 1.0)
+gen.add("rightbottom_threshold", double_t, 0, "Rightbottom threshold", 0.7, 0.5, 1.0)
+gen.add("up_threshold", double_t, 0, "Up threshold", 0.7, 0.5, 1.0)
+
+exit(gen.generate(PACKAGE, "feedback_cvsa", "CVSA"))
